+++ test/common/inspector-helper.js
