+++ test/fixtures/inspector-global-function.js
