--- test/inspector/inspector.status
@@ -1,16 +1,0 @@
-prefix inspector
-
-# To mark a test as flaky, list the test name in the appropriate section
-# below, without ".js", followed by ": PASS,FLAKY". Example:
-# sample-test                        : PASS,FLAKY
-
-[true] # This section applies to all platforms
-
-[$system==aix]
-test-stop-profile-after-done: PASS, FLAKY
-
-[$system==win32]
-test-bindings               : PASS, FLAKY
-test-debug-end              : PASS, FLAKY
-test-stop-profile-after-done: PASS, FLAKY
-
