+++ test/parallel/parallel.status
@@ -7,6 +7,9 @@ prefix parallel
 [true] # This section applies to all platforms
 
 [$system==win32]
+test-inspector-bindings               : PASS, FLAKY
+test-inspector-debug-end              : PASS, FLAKY
+test-inspector-stop-profile-after-done: PASS, FLAKY
 
 [$system==linux]
 
@@ -19,3 +22,5 @@ test-npm-install:      PASS,FLAKY
 
 [$system==freebsd]
 
+[$system==aix]
+test-inspector-stop-profile-after-done: PASS, FLAKY
