+++ test/parallel/test-inspector-async-hook-setup-at-signal.js
@@ -3,7 +3,7 @@ const common = require('../common');
 common.skipIfInspectorDisabled();
 common.skipIf32Bits();
 common.crashOnUnhandledRejection();
-const { NodeInstance } = require('../inspector/inspector-helper.js');
+const { NodeInstance } = require('../common/inspector-helper.js');
 const assert = require('assert');
 
 const script = `
