+++ test/parallel/test-inspector-async-stack-traces-promise-then.js
@@ -3,7 +3,7 @@ const common = require('../common');
 common.skipIfInspectorDisabled();
 common.skipIf32Bits();
 common.crashOnUnhandledRejection();
-const { NodeInstance } = require('./inspector-helper');
+const { NodeInstance } = require('../common/inspector-helper');
 const assert = require('assert');
 
 const script = `runTest();
