+++ test/parallel/test-inspector-async-stack-traces-set-interval.js
@@ -3,24 +3,20 @@ const common = require('../common');
 common.skipIfInspectorDisabled();
 common.skipIf32Bits();
 common.crashOnUnhandledRejection();
-const { NodeInstance } = require('./inspector-helper.js');
+const { NodeInstance } = require('../common/inspector-helper');
 const assert = require('assert');
 
-const script = `
-setTimeout(() => {
-  debugger;
-  process.exitCode = 55;
-}, 50);
-`;
+const script = 'setInterval(() => { debugger; }, 50);';
 
-async function skipBreakpointAtStart(session) {
+async function skipFirstBreakpoint(session) {
+  console.log('[test]', 'Skipping the first breakpoint in the eval script');
   await session.waitForBreakOnLine(0, '[eval]');
   await session.send({ 'method': 'Debugger.resume' });
 }
 
 async function checkAsyncStackTrace(session) {
   console.error('[test]', 'Verify basic properties of asyncStackTrace');
-  const paused = await session.waitForBreakOnLine(4, '[eval]');
+  const paused = await session.waitForBreakOnLine(2, '[eval]');
   assert(paused.params.asyncStackTrace,
          `${Object.keys(paused.params)} contains "asyncStackTrace" property`);
   assert(paused.params.asyncStackTrace.description, 'Timeout');
@@ -40,11 +36,13 @@ async function runTests() {
       'params': { 'patterns': [] } },
     { 'method': 'Runtime.runIfWaitingForDebugger' }
   ]);
-  await skipBreakpointAtStart(session);
+
+  await skipFirstBreakpoint(session);
   await checkAsyncStackTrace(session);
 
-  await session.runToCompletion();
-  assert.strictEqual(55, (await instance.expectShutdown()).exitCode);
+  console.error('[test]', 'Stopping child instance');
+  session.disconnect();
+  instance.kill();
 }
 
 runTests();
