+++ test/parallel/test-inspector-bindings.js
