+++ test/parallel/test-inspector-break-e.js
@@ -4,7 +4,7 @@ const common = require('../common');
 common.skipIfInspectorDisabled();
 
 const assert = require('assert');
-const { NodeInstance } = require('./inspector-helper.js');
+const { NodeInstance } = require('../common/inspector-helper.js');
 
 async function runTests() {
   const instance = new NodeInstance(undefined, 'console.log(10)');
