+++ test/parallel/test-inspector-debug-brk-flag.js
@@ -1,17 +1,13 @@
 'use strict';
 const common = require('../common');
-const fixtures = require('../common/fixtures');
 
 common.skipIfInspectorDisabled();
 
 const assert = require('assert');
-const { NodeInstance } = require('./inspector-helper.js');
-
-const script = fixtures.path('throws_error.js');
+const { mainScriptPath,
+        NodeInstance } = require('../common/inspector-helper.js');
 
 async function testBreakpointOnStart(session) {
-  console.log('[test]',
-              'Verifying debugger stops on start (--inspect-brk option)');
   const commands = [
     { 'method': 'Runtime.enable' },
     { 'method': 'Debugger.enable' },
@@ -27,19 +23,19 @@ async function testBreakpointOnStart(session) {
     { 'method': 'Runtime.runIfWaitingForDebugger' }
   ];
 
-  await session.send(commands);
-  await session.waitForBreakOnLine(0, script);
+  session.send(commands);
+  await session.waitForBreakOnLine(0, mainScriptPath);
 }
 
-
-async function runTest() {
-  const child = new NodeInstance(undefined, undefined, script);
+async function runTests() {
+  const child = new NodeInstance(['--inspect', '--debug-brk']);
   const session = await child.connectInspectorSession();
+
   await testBreakpointOnStart(session);
   await session.runToCompletion();
-  assert.strictEqual(1, (await child.expectShutdown()).exitCode);
+
+  assert.strictEqual(55, (await child.expectShutdown()).exitCode);
 }
 
 common.crashOnUnhandledRejection();
-
-runTest();
+runTests();
