+++ test/parallel/test-inspector-exception.js
@@ -1,13 +1,17 @@
 'use strict';
 const common = require('../common');
+const fixtures = require('../common/fixtures');
 
 common.skipIfInspectorDisabled();
 
 const assert = require('assert');
-const { mainScriptPath,
-        NodeInstance } = require('./inspector-helper.js');
+const { NodeInstance } = require('../common/inspector-helper.js');
+
+const script = fixtures.path('throws_error.js');
 
 async function testBreakpointOnStart(session) {
+  console.log('[test]',
+              'Verifying debugger stops on start (--inspect-brk option)');
   const commands = [
     { 'method': 'Runtime.enable' },
     { 'method': 'Debugger.enable' },
@@ -23,19 +27,19 @@ async function testBreakpointOnStart(session) {
     { 'method': 'Runtime.runIfWaitingForDebugger' }
   ];
 
-  session.send(commands);
-  await session.waitForBreakOnLine(0, mainScriptPath);
+  await session.send(commands);
+  await session.waitForBreakOnLine(0, script);
 }
 
-async function runTests() {
-  const child = new NodeInstance(['--inspect', '--debug-brk']);
-  const session = await child.connectInspectorSession();
 
+async function runTest() {
+  const child = new NodeInstance(undefined, undefined, script);
+  const session = await child.connectInspectorSession();
   await testBreakpointOnStart(session);
   await session.runToCompletion();
-
-  assert.strictEqual(55, (await child.expectShutdown()).exitCode);
+  assert.strictEqual(1, (await child.expectShutdown()).exitCode);
 }
 
 common.crashOnUnhandledRejection();
-runTests();
+
+runTest();
