+++ test/parallel/test-inspector-stops-no-file.js
