+++ test/sequential/test-inspector-enabled.js
